feat: add layout for keyboard and textarea
