+++ src/styles/style.scss
@@ -3,6 +3,7 @@
 
 body {
   font-family: 'Roboto', sans-serif;
+  background-color: #e7e7e7;
 }
 
 *,
@@ -12,51 +13,34 @@ body {
   margin: 0;
   padding: 0;
 }
-
 .app {
+  max-width: 1000px;
   min-height: 100vh;
+  margin: 0 auto;
   display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
-  background-color: #e7e7e7;
+  flex-direction: column;
 }
 
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  max-width: 800px;
+.textarea {
   width: 100%;
-
-  .title {
-    margin: 0;
-    padding: 0;
-  }
-
-  .descr {
-    font-size: 18px;
-    margin-bottom: 20px;
-  }
-
-  .input {
-    padding: 10px 5px;
-    max-width: 800px;
-    width: 100%;
-    margin: 0 auto;
-    resize: none;
-    font-size: 25px;
-  }
+  height: 200px;
+  font-size: 22px;
+  resize: none;
+  border: 1px solid #333;
+  padding: 5px 10px;
 }
 
 .keyboard {
-  margin: 20px;
-}
+  width: 100%;
+  margin-top: 20px;
 
-.row {
-  display: flex;
-  justify-content: space-between;
+  .row {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+  }
 }
 
 .key {
@@ -82,12 +66,20 @@ body {
   &.active {
     border-radius: 50px;
   }
+}
 
-  &--long {
-    width: 110px;
-  }
+.Backspace,
+.CapsLock,
+.Enter,
+.ShiftLeft,
+.ShiftRight {
+  width: 110px;
+}
 
-  &--very-long {
-    width: 410px;
-  }
+.Space {
+  width: 410px;
+}
+
+.hidden {
+  display: none;
 }
