--- src/utils/EnKeysMap.js
@@ -1,388 +1,0 @@
-const enLangKeys = [
-  {
-    caseDown: '`',
-    caseUp: '~',
-    caps: '`',
-    shiftCaps: '~',
-  },
-  {
-    caseDown: '1',
-    caseUp: '!',
-    caps: '1',
-    shiftCaps: '!',
-  },
-  {
-    caseDown: '2',
-    caseUp: '@',
-    caps: '2',
-    shiftCaps: '@',
-  },
-  {
-    caseDown: '3',
-    caseUp: '#',
-    caps: '3',
-    shiftCaps: '#',
-  },
-  {
-    caseDown: '4',
-    caseUp: '$',
-    caps: '4',
-    shiftCaps: '$',
-  },
-  {
-    caseDown: '5',
-    caseUp: '%',
-    caps: '5',
-    shiftCaps: '%',
-  },
-  {
-    caseDown: '6',
-    caseUp: '^',
-    caps: '6',
-    shiftCaps: '^',
-  },
-  {
-    caseDown: '7',
-    caseUp: '&',
-    caps: '7',
-    shiftCaps: '&',
-  },
-  {
-    caseDown: '8',
-    caseUp: '*',
-    caps: '8',
-    shiftCaps: '*',
-  },
-  {
-    caseDown: '9',
-    caseUp: '(',
-    caps: '9',
-    shiftCaps: '(',
-  },
-  {
-    caseDown: '0',
-    caseUp: ')',
-    caps: '0',
-    shiftCaps: ')',
-  },
-  {
-    caseDown: '-',
-    caseUp: '_',
-    caps: '-',
-    shiftCaps: '_',
-  },
-  {
-    caseDown: '=',
-    caseUp: '+',
-    caps: '=',
-    shiftCaps: '+',
-  },
-  {
-    caseDown: 'Backspace',
-    caseUp: 'Backspace',
-    caps: 'Backspace',
-    shiftCaps: 'Backspace',
-  },
-  {
-    caseDown: 'Tab',
-    caseUp: 'Tab',
-    caps: 'Tab',
-    shiftCaps: 'Tab',
-  },
-  {
-    caseDown: 'q',
-    caseUp: 'Q',
-    caps: 'Q',
-    shiftCaps: 'q',
-  },
-  {
-    caseDown: 'w',
-    caseUp: 'W',
-    caps: 'W',
-    shiftCaps: 'w',
-  },
-  {
-    caseDown: 'e',
-    caseUp: 'E',
-    caps: 'E',
-    shiftCaps: 'e',
-  },
-  {
-    caseDown: 'r',
-    caseUp: 'R',
-    caps: 'R',
-    shiftCaps: 'r',
-  },
-  {
-    caseDown: 't',
-    caseUp: 'T',
-    caps: 'T',
-    shiftCaps: 't',
-  },
-  {
-    caseDown: 'y',
-    caseUp: 'Y',
-    caps: 'Y',
-    shiftCaps: 'y',
-  },
-  {
-    caseDown: 'u',
-    caseUp: 'U',
-    caps: 'U',
-    shiftCaps: 'u',
-  },
-  {
-    caseDown: 'i',
-    caseUp: 'I',
-    caps: 'I',
-    shiftCaps: 'i',
-  },
-  {
-    caseDown: 'o',
-    caseUp: 'O',
-    caps: 'O',
-    shiftCaps: 'o',
-  },
-  {
-    caseDown: 'p',
-    caseUp: 'P',
-    caps: 'P',
-    shiftCaps: 'p',
-  },
-  {
-    caseDown: '[',
-    caseUp: '{',
-    caps: '[',
-    shiftCaps: '{',
-  },
-  {
-    caseDown: ']',
-    caseUp: '}',
-    caps: ']',
-    shiftCaps: '}',
-  },
-  {
-    caseDown: '\\',
-    caseUp: '|',
-    caps: '\\',
-    shiftCas: '|',
-  },
-  {
-    caseDown: 'Del',
-    caseUp: 'Del',
-    caps: 'Del',
-    shiftCaps: 'Del',
-  },
-  {
-    caseDown: 'CapsLock',
-    caseUp: 'CapsLock',
-    caps: 'CapsLock',
-    shiftCaps: 'CapsLock',
-  },
-  {
-    caseDown: 'a',
-    caseUp: 'A',
-    caps: 'A',
-    shiftCaps: 'a',
-  },
-  {
-    caseDown: 's',
-    caseUp: 'S',
-    caps: 'S',
-    shiftCaps: 's',
-  },
-  {
-    caseDown: 'd',
-    caseUp: 'D',
-    caps: 'D',
-    shiftCaps: 'd',
-  },
-  {
-    caseDown: 'f',
-    caseUp: 'F',
-    caps: 'F',
-    shiftCaps: 'f',
-  },
-  {
-    caseDown: 'g',
-    caseUp: 'G',
-    caps: 'G',
-    shiftCaps: 'g',
-  },
-  {
-    caseDown: 'h',
-    caseUp: 'H',
-    caps: 'H',
-    shiftCaps: 'h',
-  },
-  {
-    caseDown: 'j',
-    caseUp: 'J',
-    caps: 'J',
-    shiftCaps: 'j',
-  },
-  {
-    caseDown: 'k',
-    caseUp: 'K',
-    caps: 'K',
-    shiftCaps: 'k',
-  },
-  {
-    caseDown: 'l',
-    caseUp: 'L',
-    caps: 'L',
-    shiftCaps: 'l',
-  },
-  {
-    caseDown: ';',
-    caseUp: ':',
-    caps: ';',
-    shiftCaps: ':',
-  },
-  {
-    caseDown: "'",
-    caseUp: '"',
-    caps: "'",
-    shiftCaps: '"',
-  },
-  {
-    caseDown: 'Enter',
-    caseUp: 'Enter',
-    caps: 'Enter',
-    shiftCaps: 'Enter',
-  },
-  {
-    caseDown: 'Shift',
-    caseUp: 'Shift',
-    caps: 'Shift',
-    shiftCaps: 'Shift',
-  },
-  {
-    caseDown: 'z',
-    caseUp: 'Z',
-    caps: 'Z',
-    shiftCaps: 'z',
-  },
-  {
-    caseDown: 'x',
-    caseUp: 'X',
-    caps: 'X',
-    shiftCaps: 'x',
-  },
-  {
-    caseDown: 'c',
-    caseUp: 'C',
-    caps: 'C',
-    shiftCaps: 'c',
-  },
-  {
-    caseDown: 'v',
-    caseUp: 'V',
-    caps: 'V',
-    shiftCaps: 'v',
-  },
-  {
-    caseDown: 'b',
-    caseUp: 'B',
-    caps: 'B',
-    shiftCaps: 'b',
-  },
-  {
-    caseDown: 'n',
-    caseUp: 'N',
-    caps: 'N',
-    shiftCaps: 'n',
-  },
-  {
-    caseDown: 'm',
-    caseUp: 'M',
-    caps: 'M',
-    shiftCaps: 'm',
-  },
-  {
-    caseDown: ',',
-    caseUp: '<',
-    caps: ',',
-    shiftCaps: '<',
-  },
-  {
-    caseDown: '.',
-    caseUp: '>',
-    caps: '.',
-    shiftCaps: '>',
-  },
-  {
-    caseDown: '/',
-    caseUp: '?',
-    caps: '/',
-    shiftCaps: '?',
-  },
-  {
-    caseDown: '▲',
-    caseUp: '▲',
-    caps: '▲',
-    shiftCaps: '▲',
-  },
-  {
-    caseDown: 'Shift',
-    caseUp: 'Shift',
-    caps: 'Shift',
-    shiftCaps: 'Shift',
-  },
-  {
-    caseDown: 'Ctrl',
-    caseUp: 'Ctrl',
-    caps: 'Ctrl',
-    shiftCaps: 'Ctrl',
-  },
-  {
-    caseDown: 'Win',
-    caseUp: 'Win',
-    caps: 'Win',
-    shiftCaps: 'Win',
-  },
-  {
-    caseDown: 'Alt',
-    caseUp: 'Alt',
-    caps: 'Alt',
-    shiftCaps: 'Alt',
-  },
-  {
-    caseDown: ' ',
-    caseUp: ' ',
-    caps: ' ',
-    shiftCaps: ' ',
-  },
-  {
-    caseDown: 'Alt',
-    caseUp: 'Alt',
-    caps: 'Alt',
-    shiftCaps: 'Alt',
-  },
-  {
-    caseDown: '◄',
-    caseUp: '◄',
-    caps: '◄',
-    shiftCaps: '◄',
-  },
-  {
-    caseDown: '▼',
-    caseUp: '▼',
-    caps: '▼',
-    shiftCaps: '▼',
-  },
-  {
-    caseDown: '►',
-    caseUp: '►',
-    caps: '►',
-    shiftCaps: '►',
-  },
-  {
-    caseDown: 'Ctrl',
-    caseUp: 'Ctrl',
-    caps: 'Ctrl',
-    shiftCaps: 'Ctrl',
-  },
-];
-
-export default enLangKeys;
