--- src/utils/RuKeysMap.js
@@ -1,388 +1,0 @@
-const ruLangKeys = [
-  {
-    caseDown: 'ё',
-    caseUp: 'Ё',
-    caps: 'Ё',
-    shiftCaps: 'ё',
-  },
-  {
-    caseDown: '1',
-    caseUp: '!',
-    caps: '1',
-    shiftCaps: '!',
-  },
-  {
-    caseDown: '2',
-    caseUp: '""',
-    caps: '2',
-    shiftCaps: '""',
-  },
-  {
-    caseDown: '3',
-    caseUp: '№',
-    caps: '3',
-    shiftCaps: '№',
-  },
-  {
-    caseDown: '4',
-    caseUp: ';',
-    caps: '4',
-    shiftCaps: ';',
-  },
-  {
-    caseDown: '5',
-    caseUp: '%',
-    caps: '5',
-    shiftCaps: '%',
-  },
-  {
-    caseDown: '6',
-    caseUp: ':',
-    caps: '6',
-    shiftCaps: ':',
-  },
-  {
-    caseDown: '7',
-    caseUp: '?',
-    caps: '7',
-    shiftCaps: '?',
-  },
-  {
-    caseDown: '8',
-    caseUp: '*',
-    caps: '8',
-    shiftCaps: '*',
-  },
-  {
-    caseDown: '9',
-    caseUp: '(',
-    caps: '9',
-    shiftCaps: '(',
-  },
-  {
-    caseDown: '0',
-    caseUp: ')',
-    caps: '0',
-    shiftCaps: ')',
-  },
-  {
-    caseDown: '-',
-    caseUp: '_',
-    caps: '-',
-    shiftCaps: '_',
-  },
-  {
-    caseDown: '=',
-    caseUp: '+',
-    caps: '=',
-    shiftCaps: '+',
-  },
-  {
-    caseDown: 'Backspace',
-    caseUp: 'Backspace',
-    caps: 'Backspace',
-    shiftCaps: 'Backspace',
-  },
-  {
-    caseDown: 'Tab',
-    caseUp: 'Tab',
-    caps: 'Tab',
-    shiftCaps: 'Tab',
-  },
-  {
-    caseDown: 'й',
-    caseUp: 'Й',
-    caps: 'Й',
-    shiftCaps: 'й',
-  },
-  {
-    caseDown: 'ц',
-    caseUp: 'Ц',
-    caps: 'Ц',
-    shiftCaps: 'ц',
-  },
-  {
-    caseDown: 'у',
-    caseUp: 'У',
-    caps: 'У',
-    shiftCaps: 'у',
-  },
-  {
-    caseDown: 'к',
-    caseUp: 'К',
-    caps: 'К',
-    shiftCaps: 'к',
-  },
-  {
-    caseDown: 'е',
-    caseUp: 'Е',
-    caps: 'Е',
-    shiftCaps: 'е',
-  },
-  {
-    caseDown: 'н',
-    caseUp: 'Н',
-    caps: 'Н',
-    shiftCaps: 'н',
-  },
-  {
-    caseDown: 'г',
-    caseUp: 'Г',
-    caps: 'Г',
-    shiftCaps: 'г',
-  },
-  {
-    caseDown: 'ш',
-    caseUp: 'Ш',
-    caps: 'Ш',
-    shiftCaps: 'ш',
-  },
-  {
-    caseDown: 'щ',
-    caseUp: 'Щ',
-    caps: 'Щ',
-    shiftCaps: 'щ',
-  },
-  {
-    caseDown: 'з',
-    caseUp: 'З',
-    caps: 'З',
-    shiftCaps: 'з',
-  },
-  {
-    caseDown: 'х',
-    caseUp: 'Х',
-    caps: 'Х',
-    shiftCaps: 'х',
-  },
-  {
-    caseDown: 'ъ',
-    caseUp: 'Ъ',
-    caps: 'Ъ',
-    shiftCaps: 'ъ',
-  },
-  {
-    caseDown: '\\',
-    caseUp: '|',
-    caps: '\\',
-    shiftCas: '|',
-  },
-  {
-    caseDown: 'Del',
-    caseUp: 'Del',
-    caps: 'Del',
-    shiftCaps: 'Del',
-  },
-  {
-    caseDown: 'CapsLock',
-    caseUp: 'CapsLock',
-    caps: 'CapsLock',
-    shiftCaps: 'CapsLock',
-  },
-  {
-    caseDown: 'ф',
-    caseUp: 'Ф',
-    caps: 'Ф',
-    shiftCaps: 'ф',
-  },
-  {
-    caseDown: 'ы',
-    caseUp: 'Ы',
-    caps: 'Ы',
-    shiftCaps: 'ы',
-  },
-  {
-    caseDown: 'в',
-    caseUp: 'В',
-    caps: 'В',
-    shiftCaps: 'в',
-  },
-  {
-    caseDown: 'а',
-    caseUp: 'А',
-    caps: 'А',
-    shiftCaps: 'а',
-  },
-  {
-    caseDown: 'п',
-    caseUp: 'П',
-    caps: 'П',
-    shiftCaps: 'п',
-  },
-  {
-    caseDown: 'р',
-    caseUp: 'Р',
-    caps: 'Р',
-    shiftCaps: 'р',
-  },
-  {
-    caseDown: 'о',
-    caseUp: 'О',
-    caps: 'О',
-    shiftCaps: 'о',
-  },
-  {
-    caseDown: 'л',
-    caseUp: 'Л',
-    caps: 'Л',
-    shiftCaps: 'л',
-  },
-  {
-    caseDown: 'д',
-    caseUp: 'Д',
-    caps: 'Д',
-    shiftCaps: 'д',
-  },
-  {
-    caseDown: 'ж',
-    caseUp: 'Ж',
-    caps: 'Ж',
-    shiftCaps: 'ж',
-  },
-  {
-    caseDown: 'э',
-    caseUp: 'Э',
-    caps: 'Э',
-    shiftCaps: 'э',
-  },
-  {
-    caseDown: 'Enter',
-    caseUp: 'Enter',
-    caps: 'Enter',
-    shiftCaps: 'Enter',
-  },
-  {
-    caseDown: 'Shift',
-    caseUp: 'Shift',
-    caps: 'Shift',
-    shiftCaps: 'Shift',
-  },
-  {
-    caseDown: 'я',
-    caseUp: 'Я',
-    caps: 'Я',
-    shiftCaps: 'Я',
-  },
-  {
-    caseDown: 'ч',
-    caseUp: 'Ч',
-    caps: 'Ч',
-    shiftCaps: 'ч',
-  },
-  {
-    caseDown: 'с',
-    caseUp: 'С',
-    caps: 'С',
-    shiftCaps: 'с',
-  },
-  {
-    caseDown: 'м',
-    caseUp: 'М',
-    caps: 'М',
-    shiftCaps: 'м',
-  },
-  {
-    caseDown: 'и',
-    caseUp: 'И',
-    caps: 'И',
-    shiftCaps: 'и',
-  },
-  {
-    caseDown: 'т',
-    caseUp: 'Т',
-    caps: 'Т',
-    shiftCaps: 'т',
-  },
-  {
-    caseDown: 'о',
-    caseUp: 'О',
-    caps: 'О',
-    shiftCaps: 'о',
-  },
-  {
-    caseDown: 'б',
-    caseUp: 'Б',
-    caps: 'Б',
-    shiftCaps: 'б',
-  },
-  {
-    caseDown: 'ю',
-    caseUp: 'Ю',
-    caps: 'Ю',
-    shiftCaps: 'ю',
-  },
-  {
-    caseDown: '.',
-    caseUp: ',',
-    caps: '.',
-    shiftCaps: ',',
-  },
-  {
-    caseDown: '▲',
-    caseUp: '▲',
-    caps: '▲',
-    shiftCaps: '▲',
-  },
-  {
-    caseDown: 'Shift',
-    caseUp: 'Shift',
-    caps: 'Shift',
-    shiftCaps: 'Shift',
-  },
-  {
-    caseDown: 'Ctrl',
-    caseUp: 'Ctrl',
-    caps: 'Ctrl',
-    shiftCaps: 'Ctrl',
-  },
-  {
-    caseDown: 'Win',
-    caseUp: 'Win',
-    caps: 'Win',
-    shiftCaps: 'Win',
-  },
-  {
-    caseDown: 'Alt',
-    caseUp: 'Alt',
-    caps: 'Alt',
-    shiftCaps: 'Alt',
-  },
-  {
-    caseDown: ' ',
-    caseUp: ' ',
-    caps: ' ',
-    shiftCaps: ' ',
-  },
-  {
-    caseDown: 'Alt',
-    caseUp: 'Alt',
-    caps: 'Alt',
-    shiftCaps: 'Alt',
-  },
-  {
-    caseDown: '◄',
-    caseUp: '◄',
-    caps: '◄',
-    shiftCaps: '◄',
-  },
-  {
-    caseDown: '▼',
-    caseUp: '▼',
-    caps: '▼',
-    shiftCaps: '▼',
-  },
-  {
-    caseDown: '►',
-    caseUp: '►',
-    caps: '►',
-    shiftCaps: '►',
-  },
-  {
-    caseDown: 'Ctrl',
-    caseUp: 'Ctrl',
-    caps: 'Ctrl',
-    shiftCaps: 'Ctrl',
-  },
-];
-
-export default ruLangKeys;
